backend/requirements.txt
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
boto3==1.29.7
uvicorn==0.24.0
